requirements.txt
django==3.1
redis
Pillow